DASHBUILDE-88: Dataset columns are case sensitive
