--- dashbuilder-backend/dashbuilder-dataset-sql/src/test/java/org/dashbuilder/dataprovider/backend/sql/SQLDataSetTrimTest.java
@@ -1,34 +1,0 @@
-/**
- * Copyright 2014 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.dashbuilder.dataprovider.backend.sql;
-
-import org.dashbuilder.dataset.DataSetTrimTest;
-import org.junit.Test;
-
-public class SQLDataSetTrimTest extends SQLDataSetTestBase {
-
-    @Override
-    public void testAll() throws Exception {
-        testTrim();
-    }
-
-    @Test
-    public void testTrim() throws Exception {
-        DataSetTrimTest subTest = new DataSetTrimTest();
-        subTest.testTrim();
-        subTest.testDuplicatedColumns();
-    }
-}
